[pngcrush] Compute and report sum of critical chunk lengths IHDR, PLTE, IDAT,
and IEND, plus the 8-byte PNG signature instead of just the total IDAT data
length.  Simplify finding the lengths from the trial compressions, by replacing
the write function with one that simply counts the bytes that would have
been written to a trial PNG, instead of actually writing a PNG, reading it
back, and counting the IDAT bytes.  The savings, while measurable, are
not very significant -- The "-brute" and default pngcrush runs that I
tried are between one and five percent faster.  Most of the time is
expended in zlib compression of the IDATs, which is not affected by the change.